requirements.txt
torch
transformers
python-docx 